requirements.txt
google-generativeai
firebase_admin
Pillow